pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "mcp-cli"
version = "0.1.0"
description = "A cli for the Model Context Provider"
requires-python = ">=3.12"
readme = "README.md"
authors = [{ name = "Chris Hay", email = "chrishayuk@googlemail.com" }]
keywords = ['llm', 'openai', 'claude', 'mcp', 'cli']
license = { text = "MIT" }
dependencies = [
    "anyio>=4.6.2.post1",
    "asyncio>=3.4.3",
    "numpy>=2.2.3",
    "ollama>=0.4.2",
    "openai>=1.55.3",
    "pandas>=2.2.3",
    "prompt-toolkit>=3.0.50",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "typer>=0.15.2",
    "aiohttp>=3.8.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.17.0",
    "aiohttp>=3.8.0",
    "requests>=2.32.0",
]
[project.scripts]
mcp-cli = "cli.main:app"

[tool.ruff.lint]
unfixable = [
    "I",
    "F401", # Don't touch unused imports
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.3",
]
